+++ src/libopensc/log.h
@@ -64,7 +64,7 @@ void do_log2(struct sc_context *ctx, int facility, const char *file,
 	     int line, const char *func, const char *format,
 	     va_list args);
 
-void sc_hex_dump(struct sc_context *ctx, const u8 *buf, int len,
-		 char *out, int outlen);
+void sc_hex_dump(struct sc_context *ctx, const u8 *buf, size_t len,
+		 char *out, size_t outlen);
 
 #endif
